Add files via upload
The paper written to analyze the output from the ANN models. This paper was written with a deadline in mind and a program that wasn't fully functional at the time. So, the paper explains why I believe the program was failing and what I could do to make it function properly. However, the program was completed afterwards and the MSE was well within acceptable margins. I will rewrite the analysis paper at some point to fully explain the results that were seen after the program was completed and functioning properly.